+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Improved performance of connectivity calculation. Refactor of Least squares
+  gradients
+- Added Incompressible Navier Stokes PDE as a special formulation
+- Added ability for `make_nodes` function to return a dict 
+
 ### Changed
 
 ### Deprecated